requirements.txt
absl-py==2.3.1
altair==6.0.0
appnope==0.1.4
arviz==0.22.0
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dm-tree==0.1.9
executing==2.2.1
fiona==1.10.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-meridian==1.3.1
google-pasta==0.2.0
grpcio==1.76.0
h5netcdf==1.7.3
h5py==3.15.1
idna==3.11
immutabledict==4.2.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
meridian==0.4.0
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.12.0
natsort==7.1.1
nest-asyncio==1.6.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==0.5.6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
rtree==1.4.1
scipy==1.16.3
shapely==2.1.2
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.5
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
termcolor==3.2.0
tf_keras==2.18.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==2.0.1
xarray==2025.11.0
xarray-einstats==0.9.1